alarmes/python_backend/requirements.txt
tinytuya
paho-mqtt
flask
flask-cors
firebase-admin
